requirements.txt
numpy
torch
pillow
opencv-python
h5py
tqdm
decord
transformers

# 用openvla的话
pip install sentencepiece

